replacing suspendCoroutine with callbackFlow for fixing double resume problem
