+++ example/src/main/java/org/xmtp/android/example/account/WalletConnectV2Account.kt
@@ -2,15 +2,20 @@ package org.xmtp.android.example.account
 
 import android.net.Uri
 import com.walletconnect.wcmodal.client.Modal
+import kotlinx.coroutines.flow.first
 import org.web3j.crypto.Keys
-import org.xmtp.android.example.connect.DappDelegate
 import org.xmtp.android.example.connect.getPersonalSignBody
+import org.xmtp.android.example.extension.requestMethod
 import org.xmtp.android.library.SigningKey
 import org.xmtp.android.library.messages.SignatureBuilder
 import org.xmtp.proto.message.contents.SignatureOuterClass
 
 
-data class WalletConnectV2Account(val session: Modal.Model.ApprovedSession, val chain: String, private val sendSessionRequestDeepLink: (Uri) -> Unit) :
+data class WalletConnectV2Account(
+    val session: Modal.Model.ApprovedSession,
+    val chain: String,
+    private val sendSessionRequestDeepLink: (Uri) -> Unit
+) :
     SigningKey {
     override val address: String
         get() = Keys.toChecksumAddress(
@@ -35,10 +40,10 @@ data class WalletConnectV2Account(val session: Modal.Model.ApprovedSession, val 
                 chainId = "$parentChain:$chainId"
             )
         }
-
-        runCatching { DappDelegate.request(requestParams!!, sendSessionRequestDeepLink) }
+        runCatching {
+            requestMethod(requestParams!!, sendSessionRequestDeepLink).first().getOrThrow()
+        }
             .onSuccess {
-
                 return SignatureBuilder.buildFromSignatureData(it)
             }
             .onFailure {}
